split php_handler into _socket
using one file with a location in it did not work for aliases which do not reside in the webroot.
splitting that into a file without the location directive and defining that manually works fine.
for compatability and easy config php_handler is now a location block simply including the new _socket